refactor: move chat scroll to page
